countSignificantDecimals to /old-ui/app/util
