+++ old-ui/app/util.js
@@ -40,6 +40,7 @@ module.exports = {
   isHex: isHex,
   exportAsFile: exportAsFile,
   isInvalidChecksumAddress,
+  countSignificantDecimals,
 }
 
 function valuesFor (obj) {
@@ -250,4 +251,30 @@ function exportAsFile (filename, data) {
     elem.click()
     document.body.removeChild(elem)
   }
+}
+
+/**
+ * returns the length of truncated significant decimals for fiat value
+ * @param {float} val The float value to be truncated
+ * @param {number} len The length of significant decimals
+ * returns {number} The length of truncated significant decimals
+**/
+function countSignificantDecimals (val, len) {
+    if (Math.floor(val) === val) {
+      return 0
+    }
+    const decimals = val.toString().split('.')[1]
+    const decimalsArr = decimals.split('')
+    let decimalsLen = decimalsArr.slice(0).reduce((res, val, ind, arr) => {
+      if (Number(val) === 0) {
+        res += 1
+      } else {
+        arr.splice(1) // break reduce function
+      }
+      return res
+    }, 0)
+    decimalsLen += len
+    const valWithSignificantDecimals = `${Math.floor(val)}.${decimalsArr.slice(0, decimalsLen).join('')}`
+    decimalsLen = parseFloat(valWithSignificantDecimals).toString().split('.')[1].length
+    return decimalsLen || 0
 }
