+++ test/unit/app/util-test.js
@@ -1,5 +1,5 @@
 const assert = require('assert')
-const { sufficientBalance, countSignificantDecimals } = require('../../../app/scripts/lib/util')
+const { sufficientBalance } = require('../../../app/scripts/lib/util')
 
 
 describe('SufficientBalance', function () {
@@ -37,15 +37,5 @@ describe('SufficientBalance', function () {
 
     const result = sufficientBalance(tx, balance)
     assert.ok(!result, 'insufficient balance found.')
-  })
-})
-
-describe('countSignificantDecimals(val, len) function', () => {
-  it('returns correct significant decimals', () => {
-    assert.equal(6, countSignificantDecimals(0.00001232756347, 2))
-    assert.equal(4, countSignificantDecimals(0.00010000003454305430504350, 2))
-    assert.equal(0, countSignificantDecimals(1.0000, 2))
-    assert.equal(0, countSignificantDecimals(2, 2))
-    assert.equal(3, countSignificantDecimals('2.03243', 2))
   })
 })
